+++ src/lib.rs
@@ -27,5 +27,57 @@
 // All public items in this library must have documentation.
 #![warn(missing_docs)]
 
+use proc_macro::TokenStream;
+
 mod name;
 mod number;
+
+/// Generate a new type for a string
+///
+/// The `name!` macro generates a new type that is backed by a `String`. The new type implements
+/// common traits like `Display` and `From<&str>` and `From<String>`. The inner value can be
+/// accessed using the `get` method.
+///
+/// # Example
+///
+/// ```
+/// use typed_fields::name;
+///
+/// // Define a new type that is backed by a `String`
+/// name!(Login);
+///
+/// // Create a new `UserId` from a `&str`
+/// let id = Login::new("jdno");
+///
+/// // Common traits like `Display` are automatically implemented for the type
+/// println!("Login: {}", id);
+/// ```
+#[proc_macro]
+pub fn name(input: TokenStream) -> TokenStream {
+    name::name_impl(input)
+}
+
+/// Generate a new type for a number
+///
+/// The `number!` macro generates a new type that is backed by an `i64`. The new type implements
+/// common traits like `Display` and `From<i64>`. The inner value can be accessed using the `get`
+/// method.
+///
+/// # Example
+///
+/// ```
+/// use typed_fields::number;
+///
+/// // Define a new type that is backed by an `i64`
+/// number!(UserId);
+///
+/// // Create a new `UserId` from an `i64`
+/// let id = UserId::new(42);
+///
+/// // Common traits like `Display` are automatically implemented for the type
+/// println!("User ID: {}", id);
+/// ```
+#[proc_macro]
+pub fn number(input: TokenStream) -> TokenStream {
+    number::number_impl(input)
+}
