+++ src/name.rs
@@ -1,34 +1,17 @@
-/// Generate a new type for a string
-///
-/// The `name!` macro generates a new type that is backed by a `String`. The new type implements
-/// common traits like `Display` and `From<&str>` and `From<String>`. The inner value can be
-/// accessed using the `get` method.
-///
-/// # Example
-///
-/// ```
-/// use typed_fields::name;
-///
-/// // Define a new type that is backed by a `String`
-/// name!(Login);
-///
-/// // Create a new `UserId` from a `&str`
-/// let id = Login::new("jdno");
-///
-/// // Common traits like `Display` are automatically implemented for the type
-/// println!("Login: {}", id);
-/// ```
-#[macro_export]
-macro_rules! name {
-    (
-        $(#[$meta:meta])*
-        $name:ident
-    ) => {
-        $(#[$meta])*
-        #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
-        pub struct $name(String);
+use proc_macro::TokenStream;
 
-        impl $name {
+use proc_macro2::Ident;
+use quote::quote;
+use syn::parse_macro_input;
+
+pub fn name_impl(input: TokenStream) -> TokenStream {
+    let ident = parse_macro_input!(input as Ident);
+
+    let newtype = quote! {
+        #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
+        pub struct #ident(String);
+
+        impl #ident {
             pub fn new(name: impl Into<String>) -> Self {
                 Self(name.into())
             }
@@ -38,69 +21,24 @@ macro_rules! name {
             }
         }
 
-        impl std::fmt::Display for $name {
+        impl std::fmt::Display for #ident {
             fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
                 write!(f, "{}", self.0)
             }
         }
 
-        impl From<&str> for $name {
-            fn from(string: &str) -> $name {
-                $name::new(string)
+        impl From<&str> for #ident {
+            fn from(string: &str) -> #ident {
+                #ident::new(string)
             }
         }
 
-        impl From<String> for $name {
-            fn from(string: String) -> $name {
-                $name::new(string)
+        impl From<String> for #ident {
+            fn from(string: String) -> #ident {
+                #ident::new(string)
             }
         }
     };
-}
 
-#[cfg(test)]
-mod tests {
-    name!(TestName);
-
-    #[test]
-    fn get() {
-        let name = TestName::new("test");
-
-        assert_eq!("test", name.get());
-    }
-
-    #[test]
-    fn trait_display() {
-        let name = TestName::new("test");
-
-        assert_eq!("test", name.to_string());
-    }
-
-    #[test]
-    fn trait_from_string() {
-        let _name: TestName = String::from("test").into();
-    }
-
-    #[test]
-    fn trait_from_str() {
-        let _name: TestName = "test".into();
-    }
-
-    #[test]
-    fn trait_send() {
-        fn assert_send<T: Send>() {}
-        assert_send::<TestName>();
-    }
-
-    #[test]
-    fn trait_sync() {
-        fn assert_sync<T: Sync>() {}
-        assert_sync::<TestName>();
-    }
-
-    #[test]
-    fn trait_unpin() {
-        fn assert_unpin<T: Unpin>() {}
-        assert_unpin::<TestName>();
-    }
+    newtype.into()
 }
